# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CO-PO/PSO Attainment Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: departmentId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── DEPARTMENTS ─────────────────────────────────────────────────────────────
...
  /audit-logs:
    get:
      operationId: listAuditLogs
      tags: [users]
      summary: List audit logs
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: action
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogListResponse"
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [SUPER_ADMIN, HOD, FACULTY, IQAC]
        isActive:
          type: boolean
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    UserInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: string
          enum: [SUPER_ADMIN, HOD, FACULTY, IQAC]
        departmentId:
          type: ["integer", "null"]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [SUPER_ADMIN, HOD, FACULTY, IQAC]
        isActive:
          type: boolean
        departmentId:
          type: ["integer", "null"]
    UserListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── DEPARTMENTS ───────────────────────────────────────────────────────────
...
    AuditLog:
      type: object
      required: [id, userId, userName, action, entity, entityId, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        action:
          type: string
        entity:
          type: string
        entityId:
          type: ["integer", "null"]
        details:
          type: ["string", "null"]
        ipAddress:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    AuditLogListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer